Filter by order number (Task#7739).

+++ src/DriverApp/app/src/main/java/com/abona_erp/driver/app/ui/feature/main/fragment/history/HistoryFragment.java
@@ -14,6 +14,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.abona_erp.driver.app.App;
+import com.abona_erp.driver.app.BuildConfig;
 import com.abona_erp.driver.app.R;
 import com.abona_erp.driver.app.ui.event.PageEvent;
 import com.abona_erp.driver.app.ui.feature.main.PageItemDescriptor;
@@ -26,8 +27,8 @@ public class HistoryFragment extends Fragment {
   private RecyclerView recyclerView;
   private HistoryAdapter historyAdapter;
   private RecyclerView.LayoutManager layoutManager;
-  private AppCompatButton btnClearLog;
-  private int currentTaskId = 0;
+  private AppCompatButton btnClearLog, btnOrderNo, btnTaskId;
+  private int currentTaskId = 0, currentOrderNo = 0;
   public HistoryFragment() {
     // Required empty public constructor.
   }
@@ -44,9 +45,9 @@ public class HistoryFragment extends Fragment {
     View root = inflater.inflate(R.layout.fragment_history, container, false);
     if (getArguments() != null) {
       currentTaskId = getArguments().getInt(getResources().getString(R.string.key_taskId));
-      if(currentTaskId!=0) historyViewModel.setLogsWithTaskId(currentTaskId);
-      else historyViewModel.setHistoryLogs();
+      currentOrderNo = getArguments().getInt(getResources().getString(R.string.key_orderNo));
     }
+    historyViewModel.setHistoryLogs();
     initComponents(root);
     return root;
   }
@@ -67,9 +68,33 @@ public class HistoryFragment extends Fragment {
 
     recyclerView.setAdapter(historyAdapter);
 
-    historyViewModel.getHistoryLogs().observe(getViewLifecycleOwner(), logItems -> historyAdapter.swapData(logItems));
+    historyViewModel.getHistoryLogs().observe(getViewLifecycleOwner(), logItems -> historyViewModel.refreshLogs(logItems, currentTaskId, currentOrderNo));
+    historyViewModel.getFilteredLogs().observe(getViewLifecycleOwner(), logItems -> historyAdapter.swapData(logItems));
 
-    btnClearLog = root.findViewById(R.id.btn_clear_log);
-    btnClearLog.setOnClickListener(v -> historyViewModel.deleteChangeHistory());
+    if (BuildConfig.DEBUG) {
+      btnClearLog = root.findViewById(R.id.btn_clear_log);
+      btnClearLog.setOnClickListener(v -> historyViewModel.deleteChangeHistory());
+      btnClearLog.setVisibility(View.VISIBLE);
+    } else {
+      btnClearLog.setVisibility(View.GONE);
+    }
+
+    btnOrderNo = root.findViewById(R.id.btn_order_no);
+    btnOrderNo.setOnClickListener(v -> {
+      refreshByOrderNo();
+    });
+
+    btnTaskId = root.findViewById(R.id.btn_task_id);
+    btnTaskId.setOnClickListener(v -> {
+      refreshByTaskId();
+    });
+  }
+
+  private void refreshByOrderNo() {
+    if (currentOrderNo != 0) historyViewModel.setHistoryLogsByOrderNumber();
+  }
+
+  private void refreshByTaskId() {
+    if (currentTaskId != 0) historyViewModel.setLogsWithTaskId();
   }
 }
